3_Spring_MVC2/login/src/test/java/hello/login/web/session/SessionManagerTest.java
package hello.login.web.session;

import hello.login.domain.member.Member;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.springframework.mock.http.server.reactive.MockServerHttpResponse;
import org.springframework.mock.web.MockHttpServletRequest;
import org.springframework.mock.web.MockHttpServletResponse;
import org.springframework.test.web.servlet.MockHttpServletRequestDsl;

import javax.servlet.http.HttpServletResponse;

import static org.assertj.core.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.*;

class SessionManagerTest {
    SessionManager sessionManager = new SessionManager();

    @Test
    void SessionTest() {
        MockHttpServletResponse response = new MockHttpServletResponse();
        //세션 생성.(서버의 응답)
        Member member = new Member();
        sessionManager.createSession(member, response); // 해당 맴버에 쿠키만들어서 response에 담아줌.


        // 요청에 응답쿠키 저장.(클라의 요청)
        MockHttpServletRequest request = new MockHttpServletRequest();
        request.setCookies(response.getCookies()); // mysessionID=kdjcsacfkjb

        // 세션 조회
        Object result = sessionManager.getSession(request);
        assertThat(result).isEqualTo(member);

        // 세션만료
        sessionManager.expire(request);
        Object expired = sessionManager.getSession(request);
        assertThat(expired);
    }
}